+++ pyros/rosinterface/api/__init__.py
@@ -1,0 +1,7 @@
+# CORE ROS API
+# This is useful to quickly patch ROS problems
+#
+# Note : pyros_setup shouldn't be used here, since we need it for other packages installed in ROS environment
+# and not only for ROS interface
+#
+# TODO : move this to pyros_utils after enough time has passed to make sure it s solid enough...
