+++ pyros/rosinterface/param_if_pool.py
@@ -1,34 +1,15 @@
 from __future__ import absolute_import
 from __future__ import print_function
 
-import os
-from collections import namedtuple, MutableMapping
-from copy import deepcopy, copy
-from itertools import ifilter
 import logging
 
-import pyros_utils
-
-from . import rospy_safe as rospy
-import rosservice, rostopic, rosparam
-
-import re
-import ast
-import socket
-import threading
-import Queue
-
-import time
+from .api import rospy_safe as rospy
 
 # create logger
 _logger = logging.getLogger(__name__)
 # and let it propagate to parent logger, or other handler
 # the user of pyros should configure handlers
 
-from ..baseinterface import DiffTuple
-
-from .service import ServiceBack, ServiceTuple
-from .topic import TopicBack, TopicTuple
 from .param import ParamBack, ParamTuple
 
 from ..baseinterface import TransientIfPool
