+++ pyros/rosinterface/poolparam.py
@@ -1,17 +1,11 @@
 from __future__ import absolute_import
 
-import time
-
-import roslib
-
-from . import rospy_safe as rospy
-import rosnode
-
-from importlib import import_module
 from collections import Counter
 
+from .api import rosnode_safe as rosnode
 
-from .message_conversion import get_msg, get_msg_dict, populate_instance, extract_values, FieldTypeMismatchException
+from .api import rospy_safe as rospy
+from .message_conversion import get_msg, get_msg_dict
 
 
 def get_topic_msg(topic):
