+++ pyros/rosinterface/topic.py
@@ -2,17 +2,10 @@ from __future__ import absolute_import
 
 import time
 
-import roslib
-from . import rospy_safe as rospy
-
-from importlib import import_module
-from collections import deque, OrderedDict
-
-from ..baseinterface import TransientIf
-from .message_conversion import get_msg, get_msg_dict, populate_instance, extract_values, FieldTypeMismatchException
-
+from .api import rospy_safe as rospy
 from .publisher import PublisherBack
 from .subscriber import SubscriberBack
+from ..baseinterface import TransientIf
 
 
 class TopicBackTimeout(Exception):
