+++ pyros/rosinterface/topicbase.py
@@ -1,15 +1,11 @@
 from __future__ import absolute_import
 
-import time
+from importlib import import_module
 
 import roslib
-from . import rospy_safe as rospy
 
-from importlib import import_module
-from collections import deque, OrderedDict
-
-
-from .message_conversion import get_msg, get_msg_dict, populate_instance, extract_values, FieldTypeMismatchException
+from .api import rospy_safe as rospy
+from .message_conversion import get_msg, get_msg_dict
 
 
 def get_topic_msg(topic):
